Dev txc

Former-commit-id: 5efe98e49a64bfd6b432cf35b17312ca06bdf8f1

+++ src/scene_server/topo_server/core/operation/types.go
@@ -42,8 +42,8 @@ type OpCondition struct {
 
 type InstBatchInfo struct {
 	// BatchInfo batch info
-	BatchInfo *map[int]map[string]interface{} `json:"BatchInfo"`
-	InputType string                          `json:"input_type"`
+	BatchInfo *map[int64]map[string]interface{} `json:"BatchInfo"`
+	InputType string                            `json:"input_type"`
 }
 
 // ConditionItem subcondition
